backend/ecycle/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
Django==4.2.6
django-cors-headers==4.2.0
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Fiona==1.9.4.post1
geopandas==0.14.0
idna==3.4
numpy==1.26.0
packaging==23.2
pandas==2.1.1
phonenumbers==8.13.22
Pillow==10.0.1
psycopg2==2.9.9
PyJWT==2.8.0
pyproj==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
shapely==2.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.6
